chore: #22 유통기한 계산 Long -> Integer 로 반환타입 변경

+++ backend/jaegwan/src/main/java/com/bwmanager/jaegwan/ingredient/repository/IngredientDetailCustomRepositoryImpl.java
@@ -44,7 +44,7 @@ public class IngredientDetailCustomRepositoryImpl implements IngredientDetailCus
                 .select(Projections.fields(IngredientDetailResponse.class,
                         ingredientDetail.purchaseDate,
                         ingredientDetail.amount,
-                        Expressions.numberTemplate(Long.class,
+                        Expressions.numberTemplate(Integer.class,
                                 "DATEDIFF({0}, {1})",
                                 ingredientDetail.expirationDate, LocalDateTime.now()).as("leftExpirationDay")))
                 .from(ingredientDetail)
